jumping cooldown implemented

+++ Game/MathUtil.h
@@ -50,12 +50,12 @@ namespace Nitro {
 
 	inline float RadiansToDegrees(float rd)
 	{
-		return rd * (180.0f / M_PI);
+		return rd * (180.0f / (float)M_PI);
 	}
 
 	inline float DegreesToRadians(float dg)
 	{
-		return dg * (M_PI / 180.f);
+		return dg * ((float)M_PI / 180.f);
 	}
